# api/docs/swagger.yaml
basePath: /gpuvm
definitions:
  model.ControlInfraReq:
    properties:
      action:
        description: Action is one of suspend, resume, reboot, terminate.
        enum:
        - suspend
        - resume
        - reboot
        - terminate
        example: suspend
        type: string
    required:
    - action
    type: object
  model.ControlInfraResp:
    properties:
      affected:
        items:
          type: string
        type: array
    type: object
  model.CreateInfraReq:
    properties:
      name:
        description: Name becomes the Infra ID; node names get a -N suffix.
        example: gpu-lab
        type: string
    required:
    - name
    type: object
  model.DeleteInfraResp:
    properties:
      deleted:
        items:
          type: string
        type: array
    type: object
  model.ReleaseSharedResourcesResp:
    properties:
      released:
        items:
          $ref: '#/definitions/model.ResourceResult'
        type: array
    type: object
  model.ResourceResult:
    properties:
      id:
        type: string
      message:
        type: string
    type: object
  model.SimpleMsg:
    properties:
      message:
        example: Infra not found
        type: string
    type: object
  tumblebug.CommandResult:
    properties:
      command:
        additionalProperties:
          type: string
        type: object
      err:
        type: string
      nodeId:
        type: string
      stderr:
        additionalProperties:
          type: string
        type: object
      stdout:
        additionalProperties:
          type: string
        type: object
    type: object
  tumblebug.CommandResults:
    properties:
      results:
        items:
          $ref: '#/definitions/tumblebug.CommandResult'
        type: array
    type: object
  tumblebug.InfraInfo:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      node:
        items:
          $ref: '#/definitions/tumblebug.NodeInfo'
        type: array
      status:
        type: string
      statusCount:
        $ref: '#/definitions/tumblebug.StatusCountInfo'
      targetStatus:
        type: string
    type: object
  tumblebug.NodeInfo:
    properties:
      id:
        type: string
      imageId:
        type: string
      name:
        type: string
      nodeGroupId:
        type: string
      privateIP:
        type: string
      publicIP:
        type: string
      specId:
        type: string
      sshPort:
        type: integer
      status:
        type: string
    type: object
  tumblebug.StatusCountInfo:
    properties:
      countCreating:
        type: integer
      countRunning:
        type: integer
      countTotal:
        type: integer
    type: object
info:
  contact: {}
  description: REST API for the Ubuntu GPU VM lifecycle on CB-Tumblebug.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: GPU VM API
  version: 0.1.0
paths:
  /ns/{nsId}/infra:
    get:
      description: List every Infra of the namespace with its node status
      operationId: GetAllInfra
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Infra list
          schema:
            items:
              $ref: '#/definitions/tumblebug.InfraInfo'
            type: array
        "500":
          description: Infra lookup failed
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: List GPU Infras
      tags:
      - '[Infra] GPU VM Lifecycle'
    post:
      consumes:
      - application/json
      description: Provision an Ubuntu GPU VM from the configured template. The request
        is validated first and blocks until provisioning finishes, which takes several
        minutes
      operationId: PostInfra
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Infra creation request
        in: body
        name: infraReq
        required: true
        schema:
          $ref: '#/definitions/model.CreateInfraReq'
      produces:
      - application/json
      responses:
        "200":
          description: Infra created
          schema:
            $ref: '#/definitions/tumblebug.InfraInfo'
        "400":
          description: Malformed request body
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Infra creation failed
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Create a GPU Infra
      tags:
      - '[Infra] GPU VM Lifecycle'
  /ns/{nsId}/infra/{infraId}:
    delete:
      description: Terminate every node of the Infra and remove its records. Shared
        resources are kept; release them separately once the namespace holds no Infra
      operationId: DeleteInfra
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Infra ID
        in: path
        name: infraId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Infra deleted
          schema:
            $ref: '#/definitions/model.DeleteInfraResp'
        "404":
          description: Infra not found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Infra deletion failed
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Delete a GPU Infra
      tags:
      - '[Infra] GPU VM Lifecycle'
    get:
      description: Read one Infra. Use option=accessinfo for public IP, SSH port and
        private key
      operationId: GetInfra
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Infra ID
        in: path
        name: infraId
        required: true
        type: string
      - description: View to return
        enum:
        - status
        - accessinfo
        in: query
        name: option
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Infra information
          schema:
            $ref: '#/definitions/tumblebug.InfraInfo'
        "404":
          description: Infra not found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Infra lookup failed
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Get a GPU Infra
      tags:
      - '[Infra] GPU VM Lifecycle'
  /ns/{nsId}/infra/{infraId}/control:
    post:
      consumes:
      - application/json
      description: Apply suspend, resume, reboot or terminate to every node. Terminate
        stops the nodes but keeps the records, so delete the Infra afterwards to stop
        billing
      operationId: PostInfraControl
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Infra ID
        in: path
        name: infraId
        required: true
        type: string
      - description: Lifecycle action
        in: body
        name: controlReq
        required: true
        schema:
          $ref: '#/definitions/model.ControlInfraReq'
      produces:
      - application/json
      responses:
        "200":
          description: Action applied
          schema:
            $ref: '#/definitions/model.ControlInfraResp'
        "400":
          description: Unsupported action
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "404":
          description: Infra not found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Infra control failed
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Control a GPU Infra
      tags:
      - '[Infra] GPU VM Lifecycle'
  /ns/{nsId}/infra/{infraId}/gpu:
    get:
      description: Run nvidia-smi on every node and return the reported GPU model,
        memory and driver version
      operationId: GetInfraGPU
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Infra ID
        in: path
        name: infraId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GPU report per node
          schema:
            $ref: '#/definitions/tumblebug.CommandResults'
        "404":
          description: Infra not found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: GPU report failed
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Report GPUs of an Infra
      tags:
      - '[Infra] GPU VM Lifecycle'
  /ns/{nsId}/shared-resources:
    delete:
      description: 'Delete the shared vNet, security group and SSH key of the namespace.
        Delete every Infra first: CB-Tumblebug does not cascade and the CSP rejects
        dependent resources'
      operationId: DeleteSharedResources
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Shared resources released
          schema:
            $ref: '#/definitions/model.ReleaseSharedResourcesResp'
        "500":
          description: Shared resource release failed
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Release shared resources
      tags:
      - '[Infra] GPU VM Lifecycle'
  /readyz:
    get:
      description: Report whether the service finished initialization and can accept
        traffic
      operationId: GetReadyz
      produces:
      - application/json
      responses:
        "200":
          description: Service is ready
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "503":
          description: Service is still initializing
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Check service readiness
      tags:
      - '[Health] Service Health'
swagger: "2.0"